feat: Resume audio on user interaction
Previously, the audio context was resumed and playback was attempted upon clicking the "Start" button on the landing page. This change refactors the audio context resumption logic to trigger on any user interaction with the menu. This aims to provide a more seamless audio experience by unlocking playback earlier and more flexibly.

Additionally, a minor improvement was made to the `InputSection` component to handle potential null values returned from the `map` function when processing `pairsContent`.